# 8231269: CompileTask::is_unloaded is slow due to JNIHandles