--- a/src/hotspot/share/compiler/compileTask.hpp
+++ b/src/hotspot/share/compiler/compileTask.hpp
@@ -29,6 +29,8 @@
 #include "code/nmethod.hpp"
 #include "compiler/compileLog.hpp"
 #include "memory/allocation.hpp"
+#include "oops/oopHandle.hpp"
+#include "oops/weakHandle.hpp"
 #include "utilities/xmlstream.hpp"
 
 class DirectiveSet;
@@ -84,7 +86,8 @@ class CompileTask : public CHeapObj<mtCompiler> {
   Monitor*             _lock;
   int                  _compile_id;
   Method*              _method;
-  jobject              _method_holder;
+  WeakHandle           _method_unload_blocker_weak; // oop that can be used to block unloading method
+  OopHandle            _method_unload_blocker; // oop that *is* used to block unloading method
   int                  _osr_bci;
   bool                 _is_complete;
   bool                 _is_success;
@@ -107,7 +110,8 @@ class CompileTask : public CHeapObj<mtCompiler> {
   jlong                _time_queued;  // time when task was enqueued
   jlong                _time_started; // time when compilation started
   Method*              _hot_method;   // which method actually triggered this task
-  jobject              _hot_method_holder;
+  WeakHandle           _hot_method_unload_blocker_weak; // oop that can be used to block unloading hot method
+  OopHandle            _hot_method_unload_blocker; // oop that *is* used to block unloading hot method
   int                  _hot_count;    // information about its invocation counter
   CompileReason        _compile_reason;      // more info about the task
   const char*          _failure_reason;
@@ -223,6 +227,8 @@ class CompileTask : public CHeapObj<mtCompiler> {
   size_t       arena_bytes() const               { return _arena_bytes; }
 
 private:
+  static oop get_method_unload_blocker(methodHandle method);
+
   static void  print_impl(outputStream* st, Method* method, int compile_id, int comp_level,
                                       bool is_osr_method = false, int osr_bci = -1, bool is_blocking = false,
                                       const char* msg = nullptr, bool short_form = false, bool cr = true,