WIP. Matters when there are lots of `CompileTask`-s in the queue, e.g. in Leyden.