@shipilev 2025-03-17
You can model the impact it has on Leyden-style scenarios by producing many compile tasks with `-Xcomp`. For example, on my 5950X desktop and simple "Hello World" program that involves lots of javac compilation:

```
Benchmark 1: build/linux-x86_64-server-release/images/jdk/bin/java -Xcomp -XX:TieredStopAtLevel=1 Hello.java

# Before
  Time (mean ± σ):      1.946 s ±  0.007 s    [User: 1.830 s, System: 0.151 s]
  Range (min … max):    1.932 s …  1.956 s    10 runs

# After
  Time (mean ± σ):      1.906 s ±  0.011 s    [User: 1.782 s, System: 0.159 s]
  Range (min … max):    1.890 s …  1.924 s    10 runs
```