@shipilev 2025-04-23
Pushed the `Method*` encapsulation. SA needs fixes now, but I'll test how well this works on other tests.