@shipilev 2025-04-23
> SA needs fixes now, but I'll test how well this works on other tests.

Actually, I can just purge `CompileTask.java`: https://github.com/openjdk/jdk/pull/24832

I see that async-profiler uses the `CompileTask::_method` field directly, I think to see what compiler threads are up to. So it needs to be fixed after this PR lands, and it would dereference through the newly added handle. Luckily, I think that access only happens when compilation is already running, and `method*` is guaranteed to be alive. Paging @apangin for visibility.