@shipilev 2025-04-29
@kimbarrett, is this what you had in mind here: "But I don't think that's the way to go, because I think this code shouldn't be using JNIHandles and jobjects at all. It should be using oop* from VMGlobal and VMWeak."?