@shipilev 2025-05-05
Thank you! I'll wait a bit if @kimbarrett is able to confirm this matches the idea he had back in JBS comments.