@kimbarrett 2025-07-09
> > Not sure what our opinion is w.r.t. `mutable`, but how do we feel about typing the spin lock as `mutable` and keep `is_safe()` and `method*()` const.
> 
> I like this a lot! Dropping `const` just to satisfy spin lock (an implementation detail) felt really awkward. New version uses `mutable`.

Just a drive-by reply.  `mutable` is a C++98 (and before, I think) feature, with many uses in HotSpot.  Using it here
seems fine to me.