@shipilev 2025-10-29
/open