@shipilev 2026-01-05
/open